### 2019-03-29 #1355 @whedon
[PRE REVIEW]: The MTPy software package for magnetotelluric data analysis and visualisation
**Submitting author:** @alkirkby (<a href="http://orcid.org/0000-0003-1361-440X">Alison Kirkby</a>)
**Repository:** <a href="https://github.com/MTgeophysics/mtpy" target ="_blank">https://github.com/MTgeophysics/mtpy</a>
**Version:** v0.1
