### 2019-03-29 #1355 @danielskatz
👋 @alkirkby - in your paper metadata, you say the bib file is mtpy.bib but it really is paper.bib.

Please fix this, and then enter ```@whedon generate pdf``` as a new comment to check to see that it works.
