### 2019-03-30 75939ed
update .bib file name in JOSS paper
