### 2019-03-31 #1355 @lheagy
:wave: @JKutt: would you be willing to review this submission for JOSS?
