### 2019-03-31 #1358 @lheagy
:wave: Many thanks @JKutt for being willing to review! Above there is a checklist to help guide your review. Please feel free to open issues on the [target repository](https://github.com/MTgeophysics/mtpy) with comments or suggestions for improvement in the submission and link them here by mentioning openjournals/joss-reviews#1358 in the issue text.

The author @alkirkby mentioned that she will fix the images in the paper on monday.

Please let me know if you have any questions!
