### 2019-03-31 ee262c5
add date to JOSS paper
