### 2019-03-31 43c5d49
fix links to paper figures in JOSS paper
