### 2019-04-01 MTgeophysics/mtpy#77 @JKutt
Generalizing language to include if use on a Mac is working
Suggesting the mention if only to be used explicitly on Windows or Linux.

openjournals/joss-reviews#1358
