### 2019-04-01 MTgeophysics/mtpy#78 @JKutt
Installation confusion
Found two documents on installation process, one looks used for users and the other for developer. Initially tried the development method but get errors when trying to import `mtpy.core.mt` saying GDAL is not properly configured and passing that get `AttributeError: module 'pyproj' has no attribute 'pyproj_datadir'`. The second installation from the PDF indicates Git Bash be used. Would like to avoid installing additional Git tools. A general way to edit the bashrc files would be beneficial.

openjournals/joss-reviews#1358
