### 2019-04-02 MTgeophysics/mtpy#78 @JKutt
I managed to follow the Linux installation and got a bit further on the installation.
