### 2019-04-02 MTgeophysics/mtpy#78 @zhang01GA
Thanks a lot JKutt for review and test the software. We are working to resolve the issues.

1) the developer installation guide in the README page has been removed.
2)Only the WikiPages: https://github.com/mtgeophysics/mtpy/wiki contains installation guide.
3) Which PDF did you refer to in your comments above? We guess you mean the wiki pages?
