### 2019-04-02 #1358 @JKutt
Hi @lheagy, I managed to complete my review! I had a bit of a stumble at first getting things running but after following the Linux install more thoroughly I managed to get it going. Once up and running everything went quite smooth! A lot of good stuff in here. I found a few things and created an issue for one but the others already exist (e.g. error with print statements due to python 2 vs. 3).
