### 2019-04-02 MTgeophysics/mtpy#77 @JKutt
Installation on a Mac can be accomplished by following the Linux install instructions.
