### 2019-04-02 MTgeophysics/mtpy#78 @JKutt
No problem! Thanks for the clarification. The PDF I was referring to was `MTPy User Guide.pdf`. I found it under Docs.
