### 2019-04-02 MTgeophysics/mtpy#78 @zhang01GA
The docs/MTPy User Guide.pdf  is PDF format of workshop tutorial material.
In its introduction there are two URLs pointing to a detailed online docs (module API interfaces) and the Wiki page of installation guide: A user must first go through the wiki page installation guide before doing the test run of the modules in the MTPy User Guide.pdf.

Please let me know if this is clarified now. Thanks
