### 2019-04-03 MTgeophysics/mtpy#78 @zhang01GA
Hi JButt,
We have pushed in a commit today to fix an issue related to pyproj installation version-incompatibility.
Please git pull the latest version from develop branch, to do  further testing.
The wiki page installation guide for Windows has been updated to reflect the new change. In particular the configuration part no longer need to include the export PROJ_LIB=...

Please let us know if you can install successfully and run some test scripts. For instance:
**python examples/scripts/plot_edis.py**
--
