### 2019-04-03 MTgeophysics/mtpy#78 @JKutt
> > Found two documents on installation process, one looks used for users and the other for developer. Initially tried the development method but get errors when trying to import `mtpy.core.mt` saying GDAL is not properly configured and passing that get `AttributeError: module 'pyproj' has no attribute 'pyproj_datadir'`. The second installation from the PDF indicates Git Bash be used. Would like to avoid installing additional Git tools. A general way to edit the bashrc files would be beneficial.
> > [openjournals/joss-reviews#1358](https://github.com/openjournals/joss-reviews/issues/1358)
>
> Re git bash installation in Windows. It is simply a recommendation of using mtpy easily, as it is our preference. In principle, a user does not have to install git-bash, just have python2.7/3.6 and with mtpy package + dependencies installed, then you should be able to use the mtpy modules. Hope this clarify your question.

Clarified, thanks.
