### 2019-04-03 #1358 @alkirkby
Hey @JKutt, thanks for the review! We're glad you managed to install MTPy in the end, but we are tidying up some of the issues at the moment to make it easier for users.
I was wondering if you could let me know the issue numbers for the issues you mentioned above? I can't find anything in the issues page about print statements with python 2 vs 3, for example. I need to go through the issues page as some of the issues were resolved but never closed.
