### 2019-04-04 123515e
fix plotstrike function so modal strike is correctly displayed
