### 2019-04-04 #1358 @lheagy
Many thanks for your speedy review @JKutt! @alkirkby: it looks they are issues
- MTgeophysics/mtpy#77
- MTgeophysics/mtpy#78
