### 2019-04-05 MTgeophysics/mtpy#77 @alkirkby
I've updated the link on the Wiki page to the Linux installation instructions to say that it is for Linux or Mac users
