### 2019-04-05 #1358 @alkirkby
Great thanks @lheagy. I've checked both those issues - #78 has been addressed and I have now addressed #77.
