### 2019-04-05 677772e
fixed merge error in array2raster
