### 2019-04-05 #1358 @JKutt
Hi @alkirkby the "print" statements issue was brought up in MTgeophysics/mtpy#63
