### 2019-04-08 #1358 @alkirkby
Hi @JKutt, Fei tells me that the print statements issue was resolved, are you able to specify where in MTPy you're encountering these issues?
