### 2019-04-12 #1358 @JKutt
> Hi @JKutt, Fei tells me that the print statements issue was resolved, are you able to specify where in MTPy you're encountering these issues?

@alkirkby sorry for the late reply, missed this one. I can't quite remember off the top of my head but it was in places where "print" is used instead of "print()"
