+++ mtpy/core/mt.py
@@ -11,6 +11,10 @@ import numpy as np
 import os
 import time
 import warnings
+import datetime
+import dateutil
+import xml.etree.cElementTree as ET
+from xml.dom import minidom
 
 import mtpy.core.edi as MTedi
 import mtpy.core.z as MTz
@@ -45,7 +49,25 @@ except ImportError:  # pragma: no cover
     _logger.warning('Could not find scipy.interpolate, cannot use method interpolate'
                     'check installation you can get scipy from scipy.org.')
     interp_import = False
+    
+# =============================================================================
+#  global parameters
+# =============================================================================
+dt_fmt = '%Y-%m-%dT%H:%M:%S.%f %Z'
 
+#==============================================================================
+# Need a dummy utc time zone for the date time format
+#==============================================================================
+class UTC(datetime.tzinfo):
+    """
+    An class to hold information about UTC
+    """
+    def utcoffset(self, df):
+        return datetime.timedelta(hours=0)
+    def dst(self, df):
+        return datetime.timedelta(0)
+    def tzname(self, df):
+        return "UTC"
 
 # =============================================================================
 class MT(object):
@@ -703,7 +725,7 @@ class MT(object):
         header.units = self.Site.Location.elev_units
         header.declination = self.Site.Location.declination
         header.progvers = 'MTpy'
-        header.progdate = time.strftime('%Y-%m-%d', time.gmtime())
+        header.progdate = datetime.datetime.now().isoformat()
 
         return header
 
@@ -1276,8 +1298,31 @@ class MT(object):
                 value = obj.value
 
                 setattr(self.Processing, name, value)
-
+                
     def _write_xml_file(self, xml_fn, new_Z=None, new_Tipper=None):
+        """
+        write xml file
+        """
+        
+        emtf = ET.Element('EMTF')
+        self.Site.to_xml(emtf)
+        self.FieldNotes.to_xml(emtf)
+        self.Processing.to_xml(emtf)
+        self.Provenance.to_xml(emtf)
+        self.Copyright.to_xml(emtf)
+        
+        #--> write xml file
+        xmlstr = minidom.parseString(ET.tostring(emtf)).toprettyxml(indent="   ")
+        with open(xml_fn, 'w') as fid:
+            fid.write(xmlstr)
+        
+        print('-'*72)
+        print('    Wrote xml file to: {0}'.format(xml_fn))
+        print('-'*72)
+        
+        return xml_fn
+
+    def _write_xml_file_old(self, xml_fn, new_Z=None, new_Tipper=None):
         """
         Write a xml file.
         """
@@ -1952,13 +1997,73 @@ class MT(object):
         return plot_obj
         # raise NotImplementedError
 
+# =============================================================================
+# general object 
+# =============================================================================
+class General(object):
+    """
+    General class object that will be inhereted by the other object
+    
+    Methods
+    
+    to_xml
+    to_json
+    
+    from_xml
+    from_json
+    """
+    
+    def __init__(self):
+        pass
+    
+    def to_xml(self, parent_et, exclude_none=True):
+        """
+        convert to xml format
+        """
 
-# ==============================================================================
+        for key, value in sorted(self.__dict__.items(),
+                                 key=lambda item:(item[0][0].replace('_', '') +
+                                                  item[0][1:]).lower()):
+            if value is None and exclude_none:
+                continue
+            
+            if isinstance(getattr(self, key), (Location, Instrument, Person,
+                          Software, Citation, DataQuality, FieldNotes, 
+                          Copyright, Processing)):
+                if 'electrode' in key.lower():
+                    k = 'Dipole'
+                    attr = key.split('_')[1]
+                    k_element = ET.SubElement(parent_et, k)
+                    k_element.set('name', attr.upper())
+
+                elif 'magnetometer' in key.lower():
[…]
