+++ mtpy/core/mt.py
@@ -1341,7 +1341,14 @@ class MT(object):
         
         xml_obj.Z = self.Z
         xml_obj.Tipper = self.Tipper
-
+        
+#        emtf = ET.Element('EM_TF')
+#
+#        self.Site.to_xml(emtf)
+#        self.FieldNotes.to_xml(emtf)
+#        self.Processing.to_xml(emtf)
+#        self.Provenance.to_xml(emtf)
+#        self.Copyright.to_xml(emtf)
         xml_obj = self._xml_set_provenance(xml_obj)
         xml_obj = self._xml_set_copyright(xml_obj)
         xml_obj = self._xml_set_site(xml_obj)
@@ -1350,236 +1357,6 @@ class MT(object):
         xml_obj = self._xml_set_site_layout(xml_obj)
         
         xml_obj.write_xml_file(xml_fn)
-
-    def _xml_set_site(self, xml_obj):
-        """
-        set the Site attributes in the xml object
-        """
-        xml_obj.Site.Project.value = self.Site.project
-        if type(self.Site.survey) is list:
-            xml_obj.Site.Survey.value = ','.join(self.Site.survey)
-        else:
-            xml_obj.Site.Survey.value = self.Site.survey
-        xml_obj.Site.Id.value = self.Site.id
-        xml_obj.Site.AcquiredBy.value = self.Site.acquired_by
-        xml_obj.Site.Start.value = self.Site.start_date
-        xml_obj.Site.End.value = self.Site.end_date
-        xml_obj.Site.RunList.value = self.Site.run_list
-        xml_obj.Site.Orientation.value = 'geomagnetic'
-        try:
-            xml_obj.Site.Orientation.attr = {'angle_to_geographic_north':\
-                                             '{0:.2f}'.format(self.Site.Location.declination)}
-        except ValueError:
-            xml_obj.Site.Orientation.attr = {'angle_to_geographic_north': '0.00'}
-        
-        xml_obj.Site.Location.Latitude.value = self.lat
-        xml_obj.Site.Location.Longitude.value = self.lon
-        xml_obj.Site.Location.Elevation.value = self.elev
-        xml_obj.Site.Location.Elevation.attr = {
-            'units': self.Site.Location.elev_units}
-        xml_obj.Site.Location.Declination.value = self.Site.Location.declination
-        xml_obj.Site.Location.Declination.attr = {
-            'epoch': self.Site.Location.declination_epoch}
-
-        return xml_obj
-
-    def _xml_set_field_notes(self, xml_obj):
-        """
-        Set the FieldNotes attributes of the xml object
-        """
-
-        xml_obj.FieldNotes.Instrument.Type.value = self.FieldNotes.DataLogger.type
-        xml_obj.FieldNotes.Instrument.Id.value = self.FieldNotes.DataLogger.id
-        xml_obj.FieldNotes.Instrument.Manufacturer.value = self.FieldNotes.DataLogger.manufacturer
-
-        # EX
-        xml_obj.FieldNotes.Dipole.Type.value = self.FieldNotes.Electrode_ex.type
-        xml_obj.FieldNotes.Dipole.Id.value = self.FieldNotes.Electrode_ex.id
-        xml_obj.FieldNotes.Dipole.Manufacturer.value = self.FieldNotes.Electrode_ex.manufacturer
-        xml_obj.FieldNotes.Dipole.attr = {'name': 'EX'}
-        
-        length = np.sqrt((self.FieldNotes.Electrode_ex.x2 - self.FieldNotes.Electrode_ex.x) ** 2 +
-                         (self.FieldNotes.Electrode_ex.y2 - self.FieldNotes.Electrode_ex.y) ** 2)
-        xml_obj.FieldNotes.Dipole.Length.value = length
-        try:
[…]
