### 2019-04-17 338cbdd
adjusted shapefile to normalize by phimax for each station
