### 2019-04-18 d5e46b5
some fixes to imaging.geology and modem.residual
