### 2019-04-18 #1358 @zhang01GA
Hi  @JKutt and @alkirkby:
Two weeks ago, I did fix one new print issue in mtpy/imaging/plotstrike.py
 https://github.com/MTgeophysics/mtpy/commit/c4b6f3d8ffac3636a9bb60e7fc072bffcf312f73

Today, I applied 2to3 to every *.py script in mtpy.  But had not found any print statement issue.
All active print statement has been changed to print() in the refactoring long time ago.
(some py2 prints may exist in commented out # lines. But they are not active).

Please let me know. Thanks
