### 2019-04-20 2e2dc0e
added test for no start time in edi header
