### 2019-04-25 cb788b1
added declination_epoch to mt, and added mean phase to phasetensor ellipses
