### 2019-04-26 acd4e70
minor fixes
