+++ mtpy/imaging/phase_tensor_maps.py
@@ -694,7 +694,8 @@ class PlotPhaseTensorMaps(mtpl.PlotSettings):
             newTipper = None
             fidx = 0
             if(self.interpolate):
-                newZ, newTipper = mt.interpolate([self.plot_freq],bounds_error=False)
+                newZ, newTipper = mt.interpolate([self.plot_freq],
+                                                 bounds_error=False)
             else:
                 fidx = np.argmin(np.fabs(mt.Z.freq - self.plot_freq))
 
