### 2019-05-01 #1358 @lheagy
👋 @JKutt: it looks like the box for `automated-tests` is still un-checked. Do you have any other comments wrt testing? or can this be ticked off? Thanks!
