### 2019-05-01 #1358 @JKutt
> 👋 @JKutt: it looks like the box for automated-tests is still un-checked. Do you have any other comments wrt testing? or can this be ticked off? Thanks!

@lheagy darn, I missed that one! Apologies. No concern on the testing, simply missed ticking it off. Should be good now!
