### 2019-05-01 MTgeophysics/mtpy#80 @lheagy
docstrings not showing up on website
(related to the JOSS review openjournals/joss-reviews#1358)

Currently, the docstrings describing the modules are not displaying online. For example:
https://mtpy2.readthedocs.io/en/develop/core.html
![image](https://user-images.githubusercontent.com/6361812/57027673-e5e82f00-6bf1-11e9-8ff7-f54788c32017.png)

Also, I saw that you have a user-guide in the docs directory, but this isn't easily findable from the docs website. It might be worth including a link on the main documentation page.
