### 2019-05-01 #1358 @lheagy
Many thanks @JKutt for your review!!

@alkirkby, It looks like there is one missing DOI in the paper:
```
https://doi.org/10.1111/j.1365-246x.2005.02779.x may be missing for title: Determinable and non-determinable parameters of galvanic distortion in magnetotellurics
```

Also, I posted an issue about the documentation website - currently, it looks like the auto-documentation isn't functioning as expected (MTgeophysics/mtpy#80). Please take a look when you have a chance.
