+++ paper.bib
@@ -1,5 +1,6 @@
 @article{bibbyetal2005,
 author = {Bibby, H M and Caldwell, T G and Brown, C},
+doi = {10.1111/j.1365-246x.2005.02779.x},
 file = {:M$\backslash$:/references/Bibby et al 2005 GJI Phase Tensor.pdf:pdf},
 journal = {Geophysical Journal International},
 number = {3},
