### 2019-05-01 #1358 @alkirkby
Hi @lheagy, okay great!

@zhang01GA is looking into the documentation issue and will link the user guide from the main page. I've just fixed the doi issue. Nearly there!
