### 2019-05-02 4dd5309
update plot_phase_tensor_section example
