### 2019-05-02 fe8657f
fix to ensure the print function works for both py2 and py3
