### 2019-05-02 9afd1ee
First cut code for convert modem resistivity model into geotiff
