+++ mtpy/utils/convert_modem_data_to_geogrid.py
@@ -1,0 +1,148 @@
+#! /usr/bin/env python
+"""
+Description:
+    Convert input MODEM data  and resistivity rho files into a georeferenced raster/grid format,
+    such as geotiff format, which can be visualized by GIS software.
+
+CreationDate:   1/05/2019
+Developer:      fei.zhang@ga.gov.au
+
+Revision History:
+    LastUpdate:     1/05/2019   FZ
+    LastUpdate:     dd/mm/yyyy
+"""
+
+import os,sys
+import argparse
+from pyproj import Proj
+import gdal, osr
+import numpy as np
+from mtpy.modeling.modem import Model, Data
+from mtpy.utils import gis_tools
+from mtpy.contrib.netcdf import nc
+import mtpy.contrib.netcdf.modem_to_netCDF as modem2nc
+
+
+def array2geotiff_writer(newRasterfn, rasterOrigin, pixelWidth, pixelHeight, array):
+
+    cols = array.shape[1]
+    rows = array.shape[0]
+    originX = rasterOrigin[0]
+    originY = rasterOrigin[1]
+
+    driver = gdal.GetDriverByName('GTiff')
+    outRaster = driver.Create(newRasterfn, cols, rows, 1, gdal.GDT_Byte)
+    outRaster.SetGeoTransform((originX, pixelWidth, 0, originY, 0, pixelHeight))
+    outband = outRaster.GetRasterBand(1)
+    outband.WriteArray(array)
+    outRasterSRS = osr.SpatialReference()
+    outRasterSRS.ImportFromEPSG(4326)
+    outRaster.SetProjection(outRasterSRS.ExportToWkt())
+    outband.FlushCache()
+
+
+
+def test_array2geotiff_writer(newRasterfn):
+    #rasterOrigin = (-123.25745,45.43013)
+    rasterOrigin = (149.298, -34.974)
+    pixelWidth = 0.01
+    pixelHeight = 0.01
+    # Define an array: 0 = black 1=bright
+    array = np.array([[ 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1],
+                      [ 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1],
+                      [ 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1],
+                      [ 1, 0, 1, 1, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1],
+                      [ 1, 0, 1, 0, 0, 1, 1, 0, 1, 0, 1, 0, 0, 0, 1, 0, 1, 1, 1],
+                      [ 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1],
+                      [ 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 1],
+                      [ 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1],
+                      [ 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1],
+                      [ 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1]])
+
+
+    random= np.random.rand(array.shape[0],array.shape[1])
+
+    array2= 100*array + 10*random
+    print (array2)
[…]
