### 2019-05-02 MTgeophysics/mtpy#80 @alkirkby
This has now been fixed - documentation is compiling correctly now and a link to the user guide has been added to the front page.
