### 2019-05-02 8debc3b
reverse array to make the geotiff image right onto map
