### 2019-05-03 #1358 @lheagy
Excellent, thanks @alkirkby!

Please take a look at the pdf proof below. If everything looks good, then please make a new release and archive the software on [zenodo](http://zenodo.org) or similar and ensure that the title and author list match those in the paper. Then, please post the new version number and doi here.
