### 2019-05-06 f920132
changed legend in modem_plot_response to be opaque
