### 2019-05-08 2c755ca
adding ability to read egbert outputs
