+++ mtpy/modeling/modem/plot_response.py
@@ -14,7 +14,7 @@ import numpy as np
 import os
 from matplotlib import pyplot as plt, gridspec as gridspec
 from matplotlib.ticker import MultipleLocator
-from matplotlib.ticker import FormatStrFormatter,LogFormatterSciNotation
+from matplotlib.ticker import FormatStrFormatter, LogFormatterSciNotation
 from mtpy.imaging import mtplottools as mtplottools
 from mtpy.modeling.modem.data import Data
 import sys
@@ -522,14 +522,15 @@ class PlotResponse(object):
                             marker='.')
             t_label_list = ['Re{$T_x$}', 'Im{$T_x$}', 'Re{$T_y$}', 'Im{$T_y$}']
             label_list += [['$T_{x}$'], ['$T_{y}$']]
-            for ax, label in zip(self.ax_list[-4:], t_label_list):
-                ax.legend([l1], [label], loc='upper left',
-                          markerscale=.01,
-                          borderaxespad=.05,
-                          labelspacing=.01,
-                          handletextpad=.05,
-                          borderpad=.05,
-                          prop={'size': max([self.font_size, 6])})
+            if self.plot_tipper:
+                for ax, label in zip(self.ax_list[-4:], t_label_list):
+                    ax.legend([l1], [label], loc='upper left',
+                              markerscale=.01,
+                              borderaxespad=.05,
+                              labelspacing=.01,
+                              handletextpad=.05,
+                              borderpad=.05,
+                              prop={'size': max([self.font_size, 6])})
 
 
 
@@ -556,6 +557,7 @@ class PlotResponse(object):
 
                 if aa < 4 and self.plot_z is False:
                     ax.set_yscale('log', nonposy='clip')
+                    ax.yaxis.set_major_formatter(LogFormatterSciNotation())
                     if aa == 0 or aa == 3:
                         ax.set_ylim(self.res_limits_d)
                     elif aa == 1 or aa == 2:
@@ -1698,6 +1700,7 @@ class PlotResponse(object):
                                 ax.set_ylabel('|Im[Z (mV/km nT)]|',
                                               fontdict=fontdict)
                     elif len(self.ax_list) == 4 and plot_tipper == False:
+                        ax.yaxis.set_major_formatter(LogFormatterSciNotation())
                         if self.plot_z == True:
                             ax.set_yscale('log', nonposy='clip')
                         if aa < 2:
