### 2019-05-08 9ee4cd6
adjusted how plot_topography works
