### 2019-05-10 #1358 @alkirkby
Great thanks @lheagy,
I've just created a release and archived the software: DOI: 10.5281/zenodo.2698634
We are now just awaiting final sign-off of the proof. We'll let you know when it's ready to go!
