### 2019-05-13 dcbdada
fix file open in occam2d_rewrite.py for py3 compatibility
