### 2019-05-13 e01eef2
Fix TypeError: can't multiply sequence by non-int of type 'float'
