### 2019-05-21 #1358 @alkirkby
version number is 1.0
