### 2019-05-21 #1358 @lheagy
:wave: @alkirkby, excellent! Would you mind updating the title in the zenodo archive to match the title of the paper: "The MTPy software package for magnetotelluric data analysis and visualisation". After that, we are good to go!
