### 2019-05-21 #1358 @lheagy
Many thanks @JKutt for your speedy review! and congratulations @alkirkby and team on your submission 🎉

👋 @openjournals/joss-eics, this submission is ready to be accepted!
