### 2019-05-22 #1358 @danielskatz
@alkirkby - I see a couple of reference problems that I've tried to fix via https://github.com/MTgeophysics/mtpy/pull/82

One is opening quotes that were backwards, and the other is an underscore that caused a URL to not work right.  I'm sure my fix will work for the quotes, but I'm not sure it will work for the URL. Can you accept the PR, and see if paper will compile?  If not, we can try to fix the underscore in the URL another way.
