### 2019-05-22 #1358 @alkirkby
Hi @danielskatz just wondering if the DOI issue should have resolved by now? It's still failing to load the document for me.
